Gross pitaevskii v2

+++ Julia_Code/Auxiliary functions/derivative_prod.jl
@@ -1,4 +1,4 @@
-function derivative_prod(x)
+function derivative_prod(x::Int64)
       y = 1;
       for k = 1:2:(2*x-1)
           y *= k;
